+++ fractals/src/main/java/me/catmousedog/fractals/fractals/concrete_fractals/InverseMandelbrot.java
@@ -45,7 +45,7 @@ public class InverseMandelbrot extends Fractal {
 			y = 2 * tx * y + ky;
 
 		}
-		return new FractalValue();
+		return new FractalValue(0, 0, iterations, iterations);
 	}
 
 	@Override