+++ fractals/src/main/java/me/catmousedog/fractals/fractals/concrete_fractals/JuliaShip.java
@@ -41,7 +41,7 @@ public class JuliaShip extends MouseFractal {
 			y = Math.abs(2 * tx * y) + jy;
 
 		}
-		return new FractalValue();
+		return new FractalValue(0, 0, iterations, iterations);
 	}
 
 	@Override