+++ fractals/src/main/java/me/catmousedog/fractals/ui/GUI.java
@@ -203,8 +203,8 @@ public class GUI {
 		all = new Item[] { p20, window, p10, widthjtf, p5, heightjtf, p20, location, p10, xjtf, p5, yjtf, p5, mjtf, p5,
 				njtf, p5, rjtf, p10, copypastejb, p5, locationjcb, p5, undojb, p20, calculations, p10, iterjtf, p5,
 				zoomjtf, p10, zoomjb, p10, renderjb, p5, canceljb, repaintjb, p20, fractal, p10, fractaljl, fractaljcb,
-				fractaljp, p5, functionjl, functionjcb, p5, filterjl, filterjcb, p20, picture, p10, picturesizejcb, p5,
-				picturewjtf, p5, picturehjtf, p5, picturejb, p5, picturejcb };
+				p5, fractaljp, p5, functionjl, functionjcb, p5, functionjp, p5, filterjl, filterjcb, p5, filterjp, p20,
+				picture, p10, picturesizejcb, p5, picturewjtf, p5, picturehjtf, p5, picturejb, p5, picturejcb };
 	}
 
 	/**