+++ CHANGELOG.md
@@ -278,6 +278,7 @@ The original idea of making the **JPInterface** a class to access and retrieve d
 - ordered hierarchy Fractal -> Function -> Filter
 - resources names to match new hierarchy
 - fixes to **UI** and renamed to **PanelConstruct**
+- the JPI update methods (Fractal, Function and Filter)
 
 ### Removed
 - *Filter.setFilter(Filter)* as this is done is the cloning constructor.